% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{basename_sans_ext}
\alias{basename_sans_ext}
\title{Basename without extension}
\usage{
basename_sans_ext(path)
}
\arguments{
\item{path}{a filepath}
}
\value{
basename without extension
}
\description{
Combination of [base::basename()] and [tools::file_path_sans_ext()].
}
\examples{
beter:::basename_sans_ext("folder/file.extension") # returns "file"

}
\keyword{internal}
