% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\name{beter_options}
\alias{beter_options}
\title{Set options for the beter package}
\usage{
beter_options(..., scientific)
}
\arguments{
\item{...}{positional arguments are ignored}

\item{scientific}{older version of BEAST2 do not support scientific format in the XMLs for
integer or long parameters and thus by default, scientific notation (either from toml config,
or when specified in R) is unrolled. By setting `scientific = TRUE`, scientific notation is
preserved using native R's auto-conversion (using `as.character`).}
}
\description{
Set options that affect behaviour of `beter` package.
}
