% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{find_tags}
\alias{find_tags}
\alias{find_tags_xml}
\alias{find_tags_toml}
\alias{find_tags_str}
\title{Find {{ moustache }} tags.}
\usage{
find_tags(template)

find_tags_xml(template)

find_tags_toml(template)

find_tags_str(template)
}
\arguments{
\item{template}{String that represents path to xml, toml file or text that contain tags.
For the \code{find_tags} function, if the string is a path to an existing file, first read as
using the TOML parser or as a plain text depending if the extension of file is \code{.toml}
or \code{.xml}.}
}
\value{
list of {{ moustache }} tags
}
\description{
Find {{ moustache }} tags in a string or file. \code{find_tags_str} will find tags in a string.
\code{find_tags_xml} and \code{find_tags_toml} will find tags in a file after using
an apropriate way to read it. Finally, \code{find_tags} will call above mentioned functions
depending on its input.
}
\examples{
find_tags("Foo bar {{baz}}")
}
