% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{guess_datatype}
\alias{guess_datatype}
\title{Guess datatype from sequences}
\usage{
guess_datatype(sequences)
}
\arguments{
\item{sequences}{list of strings}
}
\value{
datatype
}
\description{
Guess data type from sequences. Currently two data types are supported:
\itemize{
     \item nucleotide
     \item standard
 }
}
