% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{guess_number_of_states}
\alias{guess_number_of_states}
\title{Guess the number of states}
\usage{
guess_number_of_states(sequences)
}
\arguments{
\item{sequences}{list of strings}
}
\value{
number of states
}
\description{
Guess the number of states in sequences
}
\keyword{internal}
