% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{join}
\alias{join}
\title{join strings}
\usage{
join(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{sep}{a a character string to separate the terms.  Not "NA_character_".}

\item{collapse}{an optional character string to separate the results.  Not "NA_character_".}
}
\value{
a character vector of the concatenated values. See \code{\link{paste}}.
}
\description{
Similar as paste, but ignores NULL, NA and empty strings.
}
\examples{
beter:::join("foo", NULL, NA, "", "bar") # "foo bar"
}
