% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mkdir}
\alias{mkdir}
\title{An utility function with an effect similar to Unix mkdir -p.
Recursively create directory. If directory exists, do nothing.}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{directory}
}
\description{
An utility function with an effect similar to Unix mkdir -p.
Recursively create directory. If directory exists, do nothing.
}
\examples{
tmpdir = tempdir()
beter:::mkdir(tmpdir) # dir already exists, no error or warning is reported

}
\keyword{internal}
