% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_template.r
\name{parse_config}
\alias{parse_config}
\title{Recursively parse TOML config
First recursively parse any TOML configs linked in the template block and then process XML
chunks by substituting the {{ moustache }} tags using the content of the default block and
any parameters from the parent TOML config.}
\usage{
parse_config(file, defaults = NULL)
}
\arguments{
\item{file}{TOML config}

\item{defaults}{parameters that will replace any other parameters from the default block}
}
\value{
list of processed XML chunks from this or any daughter config
    and the default parameter block
}
\description{
There are two different ways how parameters and xml chunks sharing the same name are merged.
Parameters will overwrite parameters with the same name in their daughter config. This means
that subtemplates/subconfigs can be shared by different TOML configs and be parametrized
according to their need. On the other hand, xml chunks with the same name are merged together
as a list, this makes it possible for different subtemplates to define their own priors,
operators or logs, which can be then put in a single point without replacing each other's
structures. This makes it possible to build a BEAST2 XML.
}
