% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_template.r
\name{process_subtemplates}
\alias{process_subtemplates}
\title{Process TOML subtemplates/subconfigs in the templates block}
\usage{
process_subtemplates(subtemplates, defaults)
}
\arguments{
\item{subtemplates}{list of TOML subtemplates from the templates block}

\item{defaults}{parameters}
}
\description{
Process TOML subtemplates/subconfigs in the templates block
}
