% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.r
\name{process_xml_chunks}
\alias{process_xml_chunks}
\title{Process list of xml chunks}
\usage{
process_xml_chunks(xml_chunks, parameters)
}
\arguments{
\item{xml_chunks}{list of strings with {{ moustache }} tags}

\item{parameters}{parameters to substitute {{ moustache }} tags}
}
\value{
list of processed xml chunks
}
\description{
process xml chunks, replacing the {{ moustache }} tags with the parameters
}
