% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sequences.r
\name{read_sequences}
\alias{read_sequences}
\alias{read_fasta}
\alias{read_nexus}
\title{Read sequences from a file}
\usage{
read_sequences(file, format = NULL)

read_fasta(file)

read_nexus(file)
}
\arguments{
\item{file}{with alignment}

\item{format}{requested format}
}
\value{
named list of sequences
}
\description{
These functions read a sequence alignment file and return a named list of sequences.
Currently supported formats are `fasta` and `nexus`. Interleaved formats, where sequences are
defined in a repeating blocks, are not supported.
}
\details{
The function `[read_sequences]` tries to guess the correct format from the file extension.
Alternatively, the correct format can be specified. Currently supported formats are
\emph{fasta} and \emph{nexus}.

Functions `[read_fasta]` and `[read_nexus]` are then used to read the sequence alignment file.
}
\examples{
seq1 = c("A" = "AAA", "B" = "BBB", "C" = "CCC")
fasta = tempfile(fileext = ".fasta")
writeLines(paste0(">", names(seq1), "\n", seq1), fasta)

seq2 = read_sequences(fasta)
seq3 = read_fasta(fasta)
identical(seq1, seq2)
identical(seq1, seq3)
}
