% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sequences.r
\name{read_sequences_helper}
\alias{read_sequences_helper}
\alias{parse_nexus_header}
\alias{parse_nexus_sequences}
\title{Internal helper functions for reading alignment}
\usage{
parse_nexus_header(text)

parse_nexus_sequences(text)
}
\arguments{
\item{text}{part of sequence file parsed by a helper function}
}
\description{
These functions are used internally when alignemnt is being read and processed. They tend to
extract and proces sequences from particular part of text.
}
