% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{readalignment}
\alias{readalignment}
\alias{read_alignment}
\alias{read_fasta}
\alias{read_nexus}
\title{Read alignment}
\usage{
read_alignment(file, format = NULL)

read_fasta(file)

read_nexus(file)
}
\arguments{
\item{file}{with alignment}

\item{format}{requested format}
}
\value{
named list of sequences
}
\description{
Parse sequence alignment and return a named list of sequences.
}
\details{
The function \code{\link{read_alignment}} reads the alignment according to the format provided
with the \code{format} parameter or, if not provided, tries to guess the alignment format from
the file extension. Currently supported formats are the \emph{fasta} and \emph{nexus} formats.
Interleaved formats are not supported.
}
