% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{readalignmenthelper}
\alias{readalignmenthelper}
\alias{parse_nexus_header}
\alias{parse_nexus_sequences}
\alias{parse_fasta_sequences}
\title{Internal helper functions for reading alignment}
\usage{
parse_nexus_header(text)

parse_nexus_sequences(text)

parse_fasta_sequences(from, to, text)
}
\arguments{
\item{text}{vector of lines with sequences in particular format}

\item{from}{in a fasta file, this is a vector of begining of sequences}

\item{to}{in a fasta file, this is a vector of ends of sequences}
}
\description{
These functions are used internally when alignemnt is being read and processed. They tend to
extract and proces sequences from particular part of text.
}
