% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{settmpdir}
\alias{settmpdir}
\title{Temporarily set a path to a specific directory. Path is restored once
the calling function/frame ends.}
\usage{
settmpdir(dir)
}
\arguments{
\item{dir}{directory}
}
\description{
Temporarily set a path to a specific directory. Path is restored once
the calling function/frame ends.
}
\examples{
dir = tempdir()
test = function(){
    beter:::settmpdir(dir)
    print(getwd())
    }
test() # should print dir
print(getwd()) # back to normal

}
\keyword{internal}
