% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_utils.r
\name{sort_named_list}
\alias{sort_named_list}
\title{Sort named list}
\usage{
sort_named_list(x, decreasing = FALSE)
}
\arguments{
\item{x}{a list}

\item{decreasing}{logical. Should the sort be increasing or decreasing?}
}
\value{
sorted list
}
\description{
Reorder list according to its names.
}
\keyword{internal}
