% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config.r
\name{write_config}
\alias{write_config}
\title{Write TOML config}
\usage{
write_config(x, file)
}
\arguments{
\item{x}{a named list}

\item{file}{path to file where TOML config will be written}
}
\description{
This function will write a list as a TOML. This is not a fully fledged TOML writer and is used
only for testing purposes.
}
\examples{
config = list(
    "xml" = list("chunk" = "<xml>{{tag}}</xml>"),
    "templates" = list(),
    "defaults" = list("tag" = "Just another tag in the wall!")
     )
write_config(config, stdout())
}
