% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sequences.r
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences to a file}
\usage{
write_fasta(seq, file = "", nchar = 80)
}
\arguments{
\item{seq}{a named vector of sequences}

\item{file}{a character string naming a file}

\item{nchar}{**optional** a number of characters per line}
}
\description{
Write sequences to a file
}
\examples{
seq1 = c("A" = "AAA", "B" = "BBB", "C" = "CCC")
fasta = tempfile(fileext = ".fasta")
write_fasta(seq1, fasta)

seq2 = read_fasta(fasta)
identical(seq1, seq2)
}
