% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sequences.r
\name{write_sequences}
\alias{write_sequences}
\alias{write_fasta}
\alias{write_nexus}
\title{Write sequences to a file}
\usage{
write_sequences(seq, file, format = NULL, ...)

write_fasta(seq, file = "", nchar = 80)

write_nexus(seq, file = "", datatype = NULL)
}
\arguments{
\item{seq}{a named vector of sequences}

\item{file}{a character string naming a file}

\item{format}{**optional** write sequences in this format, currently supported
formats are `fasta` and `nexus`}

\item{...}{**optional** parameters passed to the underlying function}

\item{nchar}{**optional** a number of characters per line}

\item{datatype}{**optional** the type of data}
}
\description{
Write sequences to a file
}
\examples{
seq1 = c("A" = "AAA", "B" = "BBB", "C" = "CCC")
fasta = tempfile(fileext = ".fasta")
write_fasta(seq1, fasta)

seq2 = read_fasta(fasta)
identical(seq1, seq2)
}
